基本调好GUIButton，事件处理的部分还要改

+++ sprite/include/sprite_event.h
@@ -16,8 +16,8 @@ class SpriteFrameStartEvent : public SpriteEvent
 {
 	DeclareRTTI;
 public:
-	float m_elapsedTime;
-	SpriteFrameStartEvent(float elapsedTime)
+	double m_elapsedTime;
+	SpriteFrameStartEvent(double elapsedTime)
 		: m_elapsedTime(elapsedTime)
 	{}
 };
